requirements.txt
flask
gunicorn
tensorflow-cpu==2.12.0
numpy
opencv-python-headless
Pillow
